requirements.txt
boto>=2.38.0
botocore>=1.3.11
colorama>=0.2.5
PyYAML>=3.10
argparse>=1.2.1
